Move tests to hudi-spark-datasource
