+++ hudi-client/hudi-spark-client/src/test/java/org/apache/hudi/client/functional/TestHoodieClientOnCopyOnWriteStorage.java
@@ -62,7 +62,6 @@ import org.apache.hudi.common.model.IOType;
 import org.apache.hudi.common.model.WriteConcurrencyMode;
 import org.apache.hudi.common.model.WriteOperationType;
 import org.apache.hudi.common.table.HoodieTableMetaClient;
-import org.apache.hudi.common.table.TableSchemaResolver;
 import org.apache.hudi.common.table.marker.MarkerType;
 import org.apache.hudi.common.table.timeline.HoodieActiveTimeline;
 import org.apache.hudi.common.table.timeline.HoodieInstant;
@@ -121,7 +120,6 @@ import org.apache.hudi.testutils.HoodieClientTestBase;
 import org.apache.hudi.testutils.HoodieClientTestUtils;
 import org.apache.hudi.testutils.HoodieSparkWriteableTestTable;
 
-import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.hadoop.fs.FSDataInputStream;
 import org.apache.hadoop.fs.FileStatus;
@@ -404,39 +402,6 @@ public class TestHoodieClientOnCopyOnWriteStorage extends HoodieClientTestBase {
     insertWithConfig(config, numRecords, instant2);
     assertTrue(testTable.inflightCommitExists(instant1));
     assertTrue(testTable.commitExists(instant2));
-  }
-
-  @Test
-  public void testInertsWithEmptyCommitsHavingWriterSchemaAsNull() throws Exception {
-    HoodieWriteConfig.Builder cfgBuilder = getConfigBuilder().withAutoCommit(false);
-    addConfigsForPopulateMetaFields(cfgBuilder, true);
-    SparkRDDWriteClient client = getHoodieWriteClient(cfgBuilder.build());
-    try {
-      String firstCommit = "001";
-      int numRecords = 200;
-      JavaRDD<WriteStatus> result = insertFirstBatch(cfgBuilder.build(), client, firstCommit, "000", numRecords, SparkRDDWriteClient::insert,
-          false, false, numRecords);
-      assertTrue(client.commit(firstCommit, result), "Commit should succeed");
-
-      // Re-init client with null writer schema.
-      cfgBuilder = getConfigBuilder((String) null).withAutoCommit(false);
-      client = getHoodieWriteClient(cfgBuilder.build());
-      String secondCommit = "002";
-      client.startCommitWithTime(secondCommit);
-      JavaRDD<HoodieRecord> emptyRdd = context.emptyRDD();
-      result = client.insert(emptyRdd, secondCommit);
-      assertTrue(client.commit(secondCommit, result), "Commit should succeed");
-      try (FSDataInputStream inputStream = fs.open(testTable.getCommitFilePath(secondCommit))) {
-        String everything = FileIOUtils.readAsUTFString(inputStream);
-        HoodieCommitMetadata metadata = HoodieCommitMetadata.fromJsonString(everything, HoodieCommitMetadata.class);
-        assertTrue(metadata.getExtraMetadata().get("schema").isEmpty());
-      }
-      HoodieTableMetaClient metaClient = HoodieTableMetaClient.builder().setConf(jsc.hadoopConfiguration()).setBasePath(basePath).build();
-      TableSchemaResolver tableSchemaResolver = new TableSchemaResolver(metaClient);
-      assertEquals(Schema.parse(TRIP_EXAMPLE_SCHEMA), tableSchemaResolver.getTableAvroSchema(false));
-    } finally {
-      client.close();
-    }
   }
 
   private void insertWithConfig(HoodieWriteConfig config, int numRecords, String instant) throws Exception {
